refactored input and taskList components
